add arc-length measurement to `Path2D` with `length`, `positionAt`, `tangentAt`, & `normalAt`
- `length` is the whole-path distance (laying all the individual contours end-to-end) as an f64
- the accessors clamp their distance arg between 0 and the path's total length (but return null if the path is zero-length)
- `tangentAt` and `normalAt` return radians (so they can be passed to ctx.rotate as-is)
- `positionAt` returns an {x, y} object

+++ src/gfx/page.rs
@@ -162,7 +162,7 @@ impl PageRecorder{
       let mut wrapper = PictureRecorder::new();
       wrapper.begin_recording(self.bounds, true).draw_drawable(&mut drawable, None);
       if let Some(pict) = wrapper.finish_recording_as_picture(None){
-        self.footprint.grow(pict.approximate_bytes_used()); // update v8's accounting
+        self.footprint.grow(pict.approximate_bytes_used() as i64); // update v8's accounting
         self.layers.push(Layer::Ops(pict));
       }
       self.approx_ops = 0;
